oops fabsf doesnt exist u idiot

+++ src/ofxAnimatable.h
@@ -64,7 +64,7 @@ class ofxAnimatable{
 		bool hasFinishedAnimating();	///has the animation finished?
 		bool isWaitingForAnimationToStart();	///an animation has been scheduled with "animateToAfterDelay"
 		bool isOrWillBeAnimating();		/// object is either animating now or it's waiting to be launch animation after a delay
-		float getCurrentSpeed(){ return fabsf( direction_ * currentSpeed_ / (lastDT_ * transitionSpeed_)) ;} ///as a percentage of linear speed
+		float getCurrentSpeed(){ return fabs( direction_ * currentSpeed_ / (lastDT_ * transitionSpeed_)) ;} ///as a percentage of linear speed
 	
 		static std::string getCurveName(AnimCurve c);
